+++ codecs/vp8_packet.go
@@ -11,7 +11,7 @@ const (
 )
 
 // Payload fragments a VP8 packet across one or more byte arrays
-func (p *VP8Payloader) Payload(mtu uint16, payload []byte) [][]byte {
+func (p *VP8Payloader) Payload(mtu uint, payload []byte) [][]byte {
 	/*
 	 * https://tools.ietf.org/html/rfc7741#section-4.2
 	 *
